Implement control over used video segments
Description:
This commit introduces a new feature to the Clipster that prevents the same segment from being used more than once in the final output. 

Changes include:

1. Addition of a `used_clips` dictionary to track the segments of each video that have been used. The key is the video file path and the value is a list of tuples, each representing a used video segment.

2. Before selecting a video segment to add to the list of clips, we now check if that segment has already been used. If it has, we skip to the next iteration of the loop and try again with a new segment.

3. When a video segment is selected for use, it is added to the list of used segments in the `used_clips` dictionary.

This new feature enhances the diversity of the final video output by ensuring that the same segment is not used more than once.